requirements.txt
Flask
numpy
torch
torchvision
Pillow
opencv-python
matplotlib
shap
gunicorn
streamlit